Remove old copy-paste and fix copy-paste tests
