+++ lib/features/copy-paste/index.js
@@ -1,12 +1,12 @@
 import CopyPasteModule from 'diagram-js/lib/features/copy-paste';
 
-import NewChoreoCopyPaste from './NewChoreoCopyPaste';
+import ChoreoCopyPaste from './ChoreoCopyPaste';
 import ChoreoModdleCopy from './ChoreoModdleCopy';
 
 /**
  * We completely overwrite the standard bpmnCopyPaste and register ChoreoCopyPaste as its replacement. We needt to do this,
  * because bpmnCopyPaste registers a function in its constructor
- * EventBus.on('element.paste') that we do not want to execute an cannot otherwise 'nicely' overwrite and prevent from
+ * EventBus.on('copyPaste.copyElement') that we do not want to execute an cannot otherwise 'nicely' overwrite and prevent from
  * bpmnCopyPaste from executing on paste.
  */
 export default {
@@ -18,5 +18,5 @@ export default {
     'bpmnCopyPaste'
   ],
   moddleCopy: ['type', ChoreoModdleCopy],
-  bpmnCopyPaste: ['type', NewChoreoCopyPaste],
+  bpmnCopyPaste: ['type', ChoreoCopyPaste],
 };
